QPID-7763: Correct return type of getAllocatedDirectMemorySize from int to long
(cherry picked from commit 00a19f1953fee7c562edf11a9bb6118c253e198e)

+++ common/src/main/java/org/apache/qpid/bytebuffer/QpidByteBuffer.java
@@ -711,7 +711,7 @@ public final class QpidByteBuffer
         return _pooledBufferSize;
     }
 
-    public static int getAllocatedDirectMemorySize()
+    public static long getAllocatedDirectMemorySize()
     {
         return _pooledBufferSize * getNumberOfActivePooledBuffers();
     }
